refactoring unlock in progress 3 and signel model in main graph
