+++ android/src/main/java/io/parity/signer/screens/onboarding/UnlockScreen.kt
@@ -1,5 +1,6 @@
 package io.parity.signer.screens.onboarding
 
+import android.content.res.Configuration
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -7,8 +8,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.fragment.app.FragmentActivity
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
@@ -16,9 +17,21 @@ import io.parity.signer.R
 import io.parity.signer.components.BigButton
 import io.parity.signer.dependencygraph.ServiceLocator
 import io.parity.signer.domain.Callback
-import io.parity.signer.domain.MainFlowViewModel
 import io.parity.signer.domain.findActivity
 import io.parity.signer.ui.MainGraphRoutes
+
+
+/**
+ * Initial screen so we won't ask token/password rightaway
+ */
+fun NavGraphBuilder.unlockAppScreenFlow(globalNavController: NavHostController) {
+	composable(route = MainGraphRoutes.initialUnlockRoute) {
+		UnlockAppAuthScreen {
+			globalNavController.navigate(MainGraphRoutes.mainScreenRoute)
+		}
+	}
+}
+
 
 @Composable
 fun UnlockAppAuthScreen(onSuccess: Callback) {
@@ -36,4 +49,18 @@ fun UnlockAppAuthScreen(onSuccess: Callback) {
 		)
 		Spacer(Modifier.weight(0.5f))
 	}
+}
+
+
+@Preview(
+	name = "light", group = "themes", uiMode = Configuration.UI_MODE_NIGHT_NO,
+	showBackground = true, backgroundColor = 0xFFFFFFFF,
+)
+@Preview(
+	name = "dark", group = "themes", uiMode = Configuration.UI_MODE_NIGHT_YES,
+	showBackground = true, backgroundColor = 0xFF000000,
+)
+@Composable
+private fun PreviewUnlockAppAuthScreen() {
+	UnlockAppAuthScreen {}
 }
