+++ android/src/main/java/io/parity/signer/ui/MainScreensAppFlow.kt
@@ -7,18 +7,18 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
-import io.parity.signer.R
-import io.parity.signer.components.BigButton
 import io.parity.signer.components.panels.BottomBar
 import io.parity.signer.components.panels.TopBar
-import io.parity.signer.dependencygraph.ServiceLocator
-import io.parity.signer.domain.*
+import io.parity.signer.domain.MainFlowViewModel
+import io.parity.signer.domain.MainFlowViewModelFactory
+import io.parity.signer.domain.NavigationMigrations
+import io.parity.signer.domain.findActivity
+import io.parity.signer.screens.onboarding.UnlockAppAuthScreen
 import io.parity.signer.ui.rustnavigationselectors.*
 
 
@@ -127,20 +127,3 @@ fun SignerMainSubgraph(mainFlowViewModel: MainFlowViewModel) {
 }
 
 
-@Composable
-fun UnlockAppAuthScreen(onSuccess: Callback) {
-	val activity = LocalContext.current.findActivity() as FragmentActivity
-
-	Column(verticalArrangement = Arrangement.Center) {
-		Spacer(Modifier.weight(0.5f))
-		BigButton(
-			text = stringResource(R.string.unlock_app_button),
-			action = {
-				ServiceLocator.authentication.authenticate(activity) {
-					onSuccess()
-				}
-			}
-		)
-		Spacer(Modifier.weight(0.5f))
-	}
-}
